+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Verbose and quiet cli options added.
 - Add support for extra flags to pass to clang ([#7](https://todo.sr.ht/~mikelma/oktac/7)). 
 - New escape sequence for strings: unicode character code.
+- Support for path resolving and absolute paths when importing modules with `use`.
 
 ### Changed